requirement.txt
fastapi
uvicorn
psycopg2-binary
sentence-transformers
google-generativeai
python-dotenv
